优化

+++ src/main/java/top/longmarch/douyin/controller/DouYinVideoController.java
@@ -51,13 +51,13 @@ public class DouYinVideoController {
 
     @GetMapping("/videoList")
     public Result videoList(@RequestParam(required = false, defaultValue = DouyinParam.COUNT) Integer count,
-                            @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Integer cursor) {
+                            @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Long cursor) {
         return Result.ok().add(douYinVideoService.videoList(count, cursor));
     }
 
     @GetMapping("/videoSearch")
     public Result videoSearch(@RequestParam(required = false, defaultValue = DouyinParam.COUNT) Integer count,
-                              @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Integer cursor,
+                              @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Long cursor,
                               @RequestParam String keyword) {
         return Result.ok().add(douYinVideoService.videoSearch(count, cursor, keyword));
     }
